refactor: extract advice related classes into 'advice' package
