+++ app/src/main/java/com/codesoom/assignment/advice/NotFoundErrorAdvice.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.advice;
 
 import com.codesoom.assignment.ProductNotFoundException;
 import com.codesoom.assignment.UserNotFoundException;
